Rename job posting request object and move it under more proper folder
