+++ src/JobsInFinland.Api.Productizer/Client/IJobsInFinlandApiClient.cs
@@ -1,9 +1,9 @@
 using JobsInFinland.Api.Infrastructure.CodeGen.Model;
-using JobsInFinland.Api.Productizer.Models.Request;
+using JobsInFinland.Api.Productizer.Models.Testbed;
 
 namespace JobsInFinland.Api.Productizer.Client;
 
 public interface IJobsInFinlandApiClient
 {
-    Task<IList<Job>> GetJobsAsync(JobsRequest jobsRequest);
+    Task<IList<Job>> GetJobsAsync(JobsPostingRequest jobsPostingRequest);
 }
